Add possibility to prepend logs by a timestamp
To facilitate some time-related debugging, we noticed that knowing at
which time a log was outputed relatively to others was a very important
data point.

Even if for that particular case, we resorted to monkey-patching console
calls directly, I propose here to add an option to the logger's setLevel method,
allowing to indicate if logs should be prepended with a timestamp
(expressed as the monotonically increasing `performance.now()` value)
indicating when that log was outputed.