[Proposal] Make SharedReference's clearSignal param mandatory
We've seen minor memory leaks in the project due to forgetting to clean-up
what's basically event listeners (mainly JS-only ones, not DOM ones).

To put a supplementary level of security against memory leaks, I propose
here to make the `SharedReference` and any `IPlaybackObserver`'s
`clearSignal` parameter as mandatory.

Note that this is not always needed to clean up a listener, there's for
example the `stopListening` argument provided in the callback, "finishing"
the reference etc., so a reference without a CancellationSignal could
still not be leaking, but ensuring there's a CancellationSignal also
attached looks like a minor sacrifice to limit memory leaks to me.