Prevent logs from being outputed in our unit tests
Our unit tests had become noisy with deprecation notices and RxPlayer
log outputs.

That noise doesn't play well with how those unit tests report as we
generally configure them to have a compact output in the terminal (with
the idea that a passing test is not useful information but a breaking
test is).
Those added logs break that compact output into a unreadable mess of
uninteresting warnings, which I find especially problematic for tests.

So here, I remove the 2 culprits for those logs:

  - deprecation notices from vitest have been handled by updating the
    tests (though I'm not sure of why they print deprecation logs each
    time when the same deprecated logic is encountered multiple times,
    one signal would have been enough and preferable IMO)

  - logs tests all mock `console` functions now