tests: Remove homemade arraybuffer vitest plugin
I noticed that the arraybuffer plugin we defined to allow arraybuffer
ES imports in vitest was unnecessary.

We already start a "content server" in our integration tests to serve
contents like manifest or segments, and we could just make an HTTP
request to that content instead of importing it directly.

This seems simpler to me as we do not rely on vitest inner working and
plugin system for this: it could make it easier to debug and it
facilitates maintainance as framework updates / switching are often
un-fun timesinks.

Our vitest config file now becomes just about defining some globals,
running our content server script and launching browsers with specific
flags on listed test files, which seems simple enough to me.