Fix keySystems option changing with reuse MediaKeySystemAccess
I noticed yet another bug while continuing to add DRM integration tests
thanks to #1478.

If the `keySystems` `loadVideo` option was subtly modified between
loadVideo calls, yet if we reused the same cached `MediaKeySystemAccess`
we risked to rely on the old `keySystems` option, not on the new one.

It seems this bug was always there, so it shouldn't impact much
applications? But it seems relatively important to me.