[Proposal] Remove webdriverio deprecation notice the ugly way
Our tests run with the test framework `vitest`, which is for now
using the `webdriverio` dependency to run it on real browser
environments.

However, it seems that `vitest` is using a deprecated API for each tests
resulting in `webdriverio` outputing a warning log after each of them.

Considering that we have thousands of tests, it makes the outputs of
integration tests unreadable, output which is particularly important
when running test scripts.

This could be considered as a `vitest` issue, and they have been made
aware of it through their #6804 issue (I'm not linking it in any way
here as GitHub as the annoying tendency to spam messages to the original
issue when linked that way, but it's the one named
"`switchToFrame` deprecated in WebdriverIO v9").

But in the meantime it has been several month and it's hard to write and
run integration tests with that mess. So here I'm proposing to just
uglily patch console functions in `vitest`'s "globalSetup" script.