[MULTI_THREAD] Fix potential leak when cleaning now inexistant Period
This is a fix for a small-scale leak happening in cases where we would
be unlucky with timings on a live content in a "multithreading" mode.

It was seen while investigating all possibilities for the issue reported
by #1643, though it has an extremely low chance to actually be linked to
that issue.

This fix is about the handling of `Init`'s `"periodStreamCleaning"`
event handling.

That event allows to clean-up resources that have previously been
allocated for a specific Period in a content: track handling, boundaries
checking etc.

Previously, only in "multithreading" mode, we checked that the Period of
a `"PeriodStreamCleaning"` event was found in the Manifest before
bubbling the event up to the `API` (the RxPlayer module).

That step was unnecessary (only the `id` property of a `Period` is
needed when cleaning up) and more importantly the check could fail in
a situation where the `Period` has since been removed from a dynamic
Manifest (e.g. after playing an old `Period` which has been loaded locally
but which now does not exist anymore in the Manifest).

The only issue I can think of with that for now are relatively-low memory
leaks (which will be collected once stopping/switching content). Still.