Make our `launch_static_server` script runnable as a script
I wanted today to run an HTTPS server serving local files quickly to
test something only locally.

What I do normally was to rely on rely on the `http-server` npm package,
which may be the better way, but I always forget the syntax wanted to
provide my certificate and key file (my self-signed ones for localhost).

We also have a much simpler `launch_static_server.mjs` script which does
that, and which has the main advantage that we're more used to it, as we
use it to run tests, our demo locally and so on.

But that file was only exported as an ES6 module and not as a script.
Consequently it only worked if imported and ran from another JS file,
but not by running directly through node.js.

Moreover, I know that @FlorentBouisset had the same need two days ago.

So I thought it could in the end also run as a script. The main
""difficulty"" is here to process flags but we already do that in a
variety of other scripts, so it was quick to add.

I could do my tests.... under I'll admit a little more time than if I
just did a `npx http-server --help`, but where's the fun in that?