tests: Make Performance test report HTML and collapsed by default
Issue: performance tests results produce too much noise on PR
-------------------------------------------------------------

We recently decided to perform performance tests (which check for
performance regressions) on all Pull Requests each time a push occurs on
it.

Those performance tests post a message detailling the results when
complete. The idea is to better see the impact for each tested scenario:
a monothreaded content loading, a multithreaded one, a seek etc. -
including when tests pass.

However, we generally have very long review time in the RxPlayer team :D.
It's not unheard of to stay in review phases for weeks, months or even
sometimes years (!), time during which we continue developing/maintaining
those branches and pushing, re-triggering the tests!

The issue is not running the tests multiple times, nor it is really the
review times, the problem we now have is that after a while, we end up
with a huge part of a Pull Request summary dedicated to performance tests
results.

Current solution: HTML `<details>` element
------------------------------------------

This commit tries to reduce this issue by relying on the `<details>`
HTML element, which results in a collapsed message.

The idea is that results are still posted each time, but the table
summarizing results is now collapsed inside the comment.

To profit from this, I decided that our performance tests now produce
HTML reports instead of markdown ones. Technically, markdown specs like
CommonMark do support HTML inside, but I that point I though it would be
more compatible to just produce HTML anyway.

The inconvenients are:

  - HTML is less readable than markdown. To alleviate this, I tried to
    make HTML as readable as possible by "playing" with line breaks and
    spaces.

  - We now have functions both producing mardown tables (for stdout) and
    HTML tables (for the file report).

    This doesn't bother me too much though.